refactor(#52): User 패키지 리네이밍 및 패키지 구조 리팩토링
