+++ src/main/java/com/shiftm/shiftm/domain/member/domain/enums/Gender.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.domain.enums;
+
+public enum Gender {
+	MALE,
+	FEMALE;
+}
