+++ src/main/java/com/shiftm/shiftm/domain/member/domain/enums/Role.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.domain.enums;
+
+public enum Role {
+	USER,
+	ADMIN;
+}
