+++ src/main/java/com/shiftm/shiftm/domain/member/dto/request/EmailCodeVerificationRequest.java
@@ -1,0 +1,7 @@
+package com.shiftm.shiftm.domain.member.dto.request;
+
+public record EmailCodeVerificationRequest(
+	String email,
+	String verificationCode
+) {
+}
