+++ src/main/java/com/shiftm/shiftm/domain/member/dto/request/FindIdRequest.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.dto.request;
+
+public record FindIdRequest(
+	String email
+) {
+}
