+++ src/main/java/com/shiftm/shiftm/domain/member/dto/request/FindPasswordRequest.java
@@ -1,0 +1,7 @@
+package com.shiftm.shiftm.domain.member.dto.request;
+
+public record FindPasswordRequest(
+	String id,
+	String email
+) {
+}
