+++ src/main/java/com/shiftm/shiftm/domain/member/dto/request/IdValidationRequest.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.dto.request;
+
+public record IdValidationRequest(
+	String id
+) {
+}
