+++ src/main/java/com/shiftm/shiftm/domain/member/dto/response/EmailCodeVerificationResponse.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.dto.response;
+
+public record EmailCodeVerificationResponse(
+	boolean isVerifiedEmailCode
+) {
+}
