+++ src/main/java/com/shiftm/shiftm/domain/member/dto/response/IdValidationResponse.java
@@ -1,0 +1,6 @@
+package com.shiftm.shiftm.domain.member.dto.response;
+
+public record IdValidationResponse(
+	boolean isIdDuplicated
+) {
+}
