+++ src/main/java/com/shiftm/shiftm/domain/member/exception/IdDuplicateException.java
@@ -1,0 +1,10 @@
+package com.shiftm.shiftm.domain.member.exception;
+
+import com.shiftm.shiftm.global.error.exception.ErrorCode;
+import com.shiftm.shiftm.global.error.exception.InvalidValueException;
+
+public class IdDuplicateException extends InvalidValueException {
+	public IdDuplicateException(String id) {
+		super(id, ErrorCode.ID_DUPLICATE);
+	}
+}
