+++ src/main/java/com/shiftm/shiftm/domain/member/exception/InvalidCompanyIdException.java
@@ -1,0 +1,10 @@
+package com.shiftm.shiftm.domain.member.exception;
+
+import com.shiftm.shiftm.global.error.exception.ErrorCode;
+import com.shiftm.shiftm.global.error.exception.InvalidValueException;
+
+public class InvalidCompanyIdException extends InvalidValueException {
+	public InvalidCompanyIdException(String companyId) {
+		super(companyId, ErrorCode.INVALID_COMPANY_ID);
+	}
+}
