--- src/main/java/com/shiftm/shiftm/domain/user/domain/enums/Gender.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.domain.enums;
-
-public enum Gender {
-	MALE,
-	FEMALE;
-}
