--- src/main/java/com/shiftm/shiftm/domain/user/domain/enums/Role.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.domain.enums;
-
-public enum Role {
-	USER,
-	ADMIN;
-}
