--- src/main/java/com/shiftm/shiftm/domain/user/dto/request/EmailCodeVerificationRequest.java
@@ -1,7 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.request;
-
-public record EmailCodeVerificationRequest(
-	String email,
-	String verificationCode
-) {
-}
