--- src/main/java/com/shiftm/shiftm/domain/user/dto/request/EmailVerificationRequest.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.request;
-
-public record EmailVerificationRequest(
-	String email
-) {
-}
