--- src/main/java/com/shiftm/shiftm/domain/user/dto/request/FindPasswordRequest.java
@@ -1,7 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.request;
-
-public record FindPasswordRequest(
-	String id,
-	String email
-) {
-}
