--- src/main/java/com/shiftm/shiftm/domain/user/dto/request/IdValidationRequest.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.request;
-
-public record IdValidationRequest(
-	String id
-) {
-}
