--- src/main/java/com/shiftm/shiftm/domain/user/dto/response/EmailCodeVerificationResponse.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.response;
-
-public record EmailCodeVerificationResponse(
-	boolean isVerifiedEmailCode
-) {
-}
