--- src/main/java/com/shiftm/shiftm/domain/user/dto/response/IdValidationResponse.java
@@ -1,6 +1,0 @@
-package com.shiftm.shiftm.domain.user.dto.response;
-
-public record IdValidationResponse(
-	boolean isIdDuplicated
-) {
-}
